Language tweak to remove a small psychological barrier. (#5)
Imposter syndrome is real, _especially_ in students. I don't think that having "smart and innovative" in the introductory sentence will improve the quality of entries, but it might subtly discourage someone who has a great idea that they aren't so sure about from bothering to apply. So I would suggest that the sentence be slightly rephrased.